Split utility files
