+++ lib/PathCommand/utils/isValidPathCommandParams.ts
@@ -1,16 +1,8 @@
 import {
-  PATH_COMMAND_COMMANDS,
   PATH_COMMAND_PARAMS_MAP,
   type CommandTypes,
   type PathCommandParams,
-} from '../../'
-
-export const isValidCommandType = (
-  command: string
-): command is CommandTypes => {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  return PATH_COMMAND_COMMANDS.includes(command as any)
-}
+} from '../'
 
 export const isValidPathCommandParams = (
   command: CommandTypes,
